+++ ImmutablePendingIntent-lint/src/main/java/com/wada811/immutablependingintent/lint/ImmutablePendingIntentIssueRegistry.kt
@@ -5,7 +5,7 @@ import com.android.tools.lint.client.api.Vendor
 import com.android.tools.lint.detector.api.CURRENT_API
 import com.android.tools.lint.detector.api.Issue
 
-@Suppress("UnstableApiUsage")
+@Suppress("UnstableApiUsage", "unused")
 class ImmutablePendingIntentIssueRegistry : IssueRegistry() {
     override val issues: List<Issue> = listOf(
         UseRawPendingIntentDetector.ISSUE,
