+++ ImmutablePendingIntent/src/main/java/com/wada811/immutablependingintent/ImmutablePendingIntent.kt
@@ -1,8 +1,9 @@
 @file:JvmName("ImmutablePendingIntent")
 
-package android.app
+package com.wada811.immutablependingintent
 
 import android.annotation.SuppressLint
+import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
 import android.os.Build
